+++ DomainDrivenDesign/src/AggregateRoot.cs
@@ -1,5 +1,10 @@
 ﻿namespace FunctionalDDD.Domain;
-public abstract class AggregateRoot<TId> : Entity<TId>
+/// <summary>
+/// A DDD aggregate is a cluster of domain objects that can be treated as a single unit. 
+/// An aggregate will have one of its component objects be the aggregate root.Any references from outside the aggregate should only go to the aggregate root.
+/// The root can thus ensure the integrity of the aggregate as a whole.
+
+public abstract class AggregateRoot<TId> : Aggregate<TId>
     where TId : notnull
 {
     protected AggregateRoot(TId id) : base(id)
